Added page object for about clients section, moved test logic to page object, updated clients tests based on page object, small fixes, updated useful info
